+++ storybook/src/components/button-row/button-row.css-template.ts
@@ -6,8 +6,8 @@ export const cssButtonRow: ComponentImplementation<ButtonRow> = {
   component: "buttonRow",
   implementation: "html-css",
   template: ({ buttonTemplate, anchorTemplate }) =>
-    function buttonRowTemplate({ buttons }) {
-      return html`<div class="dso-button-row">
+    function buttonRowTemplate({ buttons, emphasized }) {
+      return html`<div class=${["dso-button-row", emphasized ? "dso-emphasized" : ""].join(" ")}>
         ${buttons.map((button) => ("variant" in button ? buttonTemplate(button) : anchorTemplate(button)))}
       </div>`;
     },